# Cython experiment to speed up GMM fitting