This should be used with caution, but can give some speedup, e.g. for resolving ~600 objects in 2D:

```
with_cython_run1 -> 7.611491999999998s
with_cython_run2 -> 7.44231s
no_cython_run1 -> 7.9705189999999995s
no_cython_run2 -> 8.174182s
```