@chaubold 2016-11-24
Okay, some more testing reveals that these timing results are not stable. So let's just discard this experiment. 

I'd say either we set up a proper C++ module for the GMM fit, or we just leave it as is.

```
Fitting gmm to a 2-merger with coords of shape (915, 3):

cython
-------
GMM fitting took 0.0182185769081 secs on average in 100 runs
GMM fitting took 0.0176032233238 secs on average in 100 runs

sklearn
-------
GMM fitting took 0.0129393935204 secs on average in 100 runs
GMM fitting took 0.0162936210632 secs on average in 100 runs
```

@JaimeIvanCervantes: On another note: for testing I used one real world merger from a 2D dataset as we see it in ilastik. Even though it is 2D we are still passing in 3 coordinates per pixel. Using only 2 might give us a little performance.